+++ familyhelper-webapi-node/src/main/java/com/dwarfeng/familyhelper/webapi/node/controller/v1/assets/ItemFileController.java
@@ -1,0 +1,291 @@
+package com.dwarfeng.familyhelper.webapi.node.controller.v1.assets;
+
+import com.dwarfeng.dutil.basic.io.IOUtil;
+import com.dwarfeng.familyhelper.assets.sdk.bean.entity.JSFixedFastJsonItemFileInfo;
+import com.dwarfeng.familyhelper.assets.stack.bean.dto.ItemFile;
+import com.dwarfeng.familyhelper.assets.stack.bean.dto.ItemFileUploadInfo;
+import com.dwarfeng.familyhelper.assets.stack.bean.entity.ItemFileInfo;
+import com.dwarfeng.familyhelper.webapi.stack.handler.system.TokenHandler;
+import com.dwarfeng.familyhelper.webapi.stack.service.assets.ItemFileResponseService;
+import com.dwarfeng.subgrade.sdk.bean.dto.FastJsonResponseData;
+import com.dwarfeng.subgrade.sdk.bean.dto.JSFixedFastJsonPagedData;
+import com.dwarfeng.subgrade.sdk.bean.dto.PagingUtil;
+import com.dwarfeng.subgrade.sdk.bean.dto.ResponseDataUtil;
+import com.dwarfeng.subgrade.sdk.bean.key.WebInputLongIdKey;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
+import com.dwarfeng.subgrade.sdk.interceptor.http.BindingCheck;
+import com.dwarfeng.subgrade.sdk.interceptor.login.LoginRequired;
+import com.dwarfeng.subgrade.stack.bean.BeanTransformer;
+import com.dwarfeng.subgrade.stack.bean.dto.PagedData;
+import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
+import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
+import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
+import com.dwarfeng.subgrade.stack.exception.ServiceExceptionMapper;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.multipart.MultipartHttpServletRequest;
+import org.springframework.web.multipart.commons.CommonsMultipartResolver;
+
+import javax.servlet.http.HttpServletRequest;
+import java.io.ByteArrayOutputStream;
+import java.io.InputStream;
+import java.nio.charset.StandardCharsets;
+import java.util.Objects;
+
+/**
+ * 项目文件控制器。
+ *
+ * @author DwArFeng
+ * @since 1.0.2
+ */
+@RestController
+@RequestMapping("/api/v1/assets")
+public class ItemFileController {
+
+    private final ItemFileResponseService service;
+
+    private final ServiceExceptionMapper sem;
+
+    private final BeanTransformer<ItemFileInfo, JSFixedFastJsonItemFileInfo> beanTransformer;
+
+    private final CommonsMultipartResolver commonsMultipartResolver;
+
+    private final TokenHandler tokenHandler;
+
+    public ItemFileController(
+            ItemFileResponseService service,
+            ServiceExceptionMapper sem,
+            BeanTransformer<ItemFileInfo, JSFixedFastJsonItemFileInfo> beanTransformer,
+            CommonsMultipartResolver commonsMultipartResolver,
+            TokenHandler tokenHandler
+    ) {
+        this.service = service;
+        this.sem = sem;
+        this.beanTransformer = beanTransformer;
+        this.commonsMultipartResolver = commonsMultipartResolver;
+        this.tokenHandler = tokenHandler;
+    }
+
+    @GetMapping("/item-file/{longId}/exists")
+    @BehaviorAnalyse
+    @LoginRequired
+    public FastJsonResponseData<Boolean> exists(HttpServletRequest request, @PathVariable("longId") Long longId) {
+        try {
+            boolean exists = service.exists(new LongIdKey(longId));
+            return FastJsonResponseData.of(ResponseDataUtil.good(exists));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(Boolean.class, e, sem));
+        }
+    }
+
+    @GetMapping("/item-file/{longId}")
+    @BehaviorAnalyse
+    @LoginRequired
+    public FastJsonResponseData<JSFixedFastJsonItemFileInfo> get(
+            HttpServletRequest request, @PathVariable("longId") Long longId
+    ) {
+        try {
+            ItemFileInfo itemFileInfo = service.get(new LongIdKey(longId));
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonItemFileInfo.of(itemFileInfo)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(JSFixedFastJsonItemFileInfo.class, e, sem));
+        }
+    }
+
+    @GetMapping("item/{itemId}/item-file/default")
+    @BehaviorAnalyse
+    @LoginRequired
+    public FastJsonResponseData<JSFixedFastJsonPagedData<JSFixedFastJsonItemFileInfo>> childForItem(
+            HttpServletRequest request,
+            @PathVariable("itemId") long itemId,
+            @RequestParam("page") int page, @RequestParam("rows") int rows
+    ) {
+        try {
+            PagedData<ItemFileInfo> childForItem = service.childForItem(
+                    new LongIdKey(itemId), new PagingInfo(page, rows)
+            );
+            PagedData<JSFixedFastJsonItemFileInfo> transform = PagingUtil.transform(
+                    childForItem, beanTransformer
+            );
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonPagedData.of(transform)));
+        } catch (Exception e) {
+            e.printStackTrace();
+            return FastJsonResponseData.of(ResponseDataUtil.bad(JSFixedFastJsonPagedData.class, e, sem));
+        }
+    }
+
+    @GetMapping("item/{itemId}/item-file/inspected-date-desc")
+    @BehaviorAnalyse
+    @LoginRequired
+    public FastJsonResponseData<JSFixedFastJsonPagedData<JSFixedFastJsonItemFileInfo>> childForItemInspectedDateDesc(
+            HttpServletRequest request,
+            @PathVariable("itemId") long itemId,
+            @RequestParam("page") int page, @RequestParam("rows") int rows
+    ) {
+        try {
+            PagedData<ItemFileInfo> childForItem = service.childForItemInspectedDateDesc(
+                    new LongIdKey(itemId), new PagingInfo(page, rows)
+            );
+            PagedData<JSFixedFastJsonItemFileInfo> transform = PagingUtil.transform(
+                    childForItem, beanTransformer
+            );
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonPagedData.of(transform)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(JSFixedFastJsonPagedData.class, e, sem));
+        }
+    }
+
+    @GetMapping("item/{itemId}/item-file/modified-date-desc")
+    @BehaviorAnalyse
+    @LoginRequired
+    public FastJsonResponseData<JSFixedFastJsonPagedData<JSFixedFastJsonItemFileInfo>> childForItemModifiedDateDesc(
+            HttpServletRequest request,
+            @PathVariable("itemId") long itemId,
+            @RequestParam("page") int page, @RequestParam("rows") int rows
+    ) {
+        try {
+            PagedData<ItemFileInfo> childForItem = service.childForItemModifiedDateDesc(
+                    new LongIdKey(itemId), new PagingInfo(page, rows)
+            );
+            PagedData<JSFixedFastJsonItemFileInfo> transform = PagingUtil.transform(
+                    childForItem, beanTransformer
+            );
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonPagedData.of(transform)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(JSFixedFastJsonPagedData.class, e, sem));
+        }
+    }
+
+    @GetMapping("item/{itemId}/item-file/origin-name-asc")
+    @BehaviorAnalyse
+    @LoginRequired
+    public FastJsonResponseData<JSFixedFastJsonPagedData<JSFixedFastJsonItemFileInfo>> childForItemOriginNameAsc(
+            HttpServletRequest request,
+            @PathVariable("itemId") long itemId,
+            @RequestParam("page") int page, @RequestParam("rows") int rows
+    ) {
+        try {
+            PagedData<ItemFileInfo> childForItem = service.childForItemOriginNameAsc(
+                    new LongIdKey(itemId), new PagingInfo(page, rows)
+            );
+            PagedData<JSFixedFastJsonItemFileInfo> transform = PagingUtil.transform(
+                    childForItem, beanTransformer
+            );
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonPagedData.of(transform)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(JSFixedFastJsonPagedData.class, e, sem));
+        }
+    }
+
+    @GetMapping("item/{itemId}/item-file/created-date-asc")
+    @BehaviorAnalyse
+    @LoginRequired
+    public FastJsonResponseData<JSFixedFastJsonPagedData<JSFixedFastJsonItemFileInfo>> childForItemCreatedDateAsc(
+            HttpServletRequest request,
+            @PathVariable("itemId") long itemId,
+            @RequestParam("page") int page, @RequestParam("rows") int rows
+    ) {
+        try {
+            PagedData<ItemFileInfo> childForItem = service.childForItemCreatedDateAsc(
+                    new LongIdKey(itemId), new PagingInfo(page, rows)
+            );
+            PagedData<JSFixedFastJsonItemFileInfo> transform = PagingUtil.transform(
+                    childForItem, beanTransformer
+            );
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonPagedData.of(transform)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(JSFixedFastJsonPagedData.class, e, sem));
+        }
+    }
+
+
+    @GetMapping("/item-file/{itemFileId}/download")
+    @BehaviorAnalyse
+    @BindingCheck
+    @LoginRequired
+    public ResponseEntity<Object> downloadItemFile(
+            HttpServletRequest request, @PathVariable("itemFileId") Long itemFileId
+    ) {
+        HttpHeaders headers = new HttpHeaders();
+        Object body;
+        try {
+            StringIdKey accountKey = tokenHandler.getAccountKey(request);
+            ItemFile itemFile = service.downloadItemFile(accountKey, new LongIdKey(itemFileId));
+            // 将文件名转换成 HTTP 标准文件名编码下的格式。
+            String fileName = adjustFileNameEncoding(itemFile.getOriginName());
+            headers.add("Content-Disposition", "attachment;filename=" + fileName);
+            body = itemFile.getContent();
+        } catch (Exception e) {
+            body = FastJsonResponseData.of(ResponseDataUtil.bad(Object.class, e, sem));
+        }
+        return new ResponseEntity<>(body, headers, HttpStatus.OK);
+    }
+
+    @SuppressWarnings("DuplicatedCode")
+    @PostMapping("/item/{itemId}/item-file/upload")
+    @BehaviorAnalyse
+    @BindingCheck
+    @LoginRequired
+    public FastJsonResponseData<Object> uploadItemFile(
+            HttpServletRequest request, @PathVariable("itemId") Long itemId
+    ) {
+        try {
+            // 通过请求解析用户。
+            StringIdKey accountKey = tokenHandler.getAccountKey(request);
+
+            // 确认请求合法。
+            if (!commonsMultipartResolver.isMultipart(request)) {
+                throw new IllegalStateException("请求不是标准的文件上传请求");
+            }
+
+            //获取 multiRequest 中的文件。
+            MultipartHttpServletRequest multipartHttpServletRequest = commonsMultipartResolver.resolveMultipart(request);
+            MultipartFile file = multipartHttpServletRequest.getFile("file");
+            if (Objects.isNull(file)) {
+                throw new IllegalStateException("请求体中缺少 file 属性");
+            }
+
+            // 解析文件内容。
+            String originFileName = file.getOriginalFilename();
+            byte[] content;
+            try (InputStream in = file.getInputStream(); ByteArrayOutputStream bout = new ByteArrayOutputStream()) {
+                IOUtil.trans(in, bout, 4096);
+                bout.flush();
+                content = bout.toByteArray();
+            }
+
+            // 将文件内容转换为接口需要的格式，并上传。
+            service.uploadItemFile(
+                    accountKey, new ItemFileUploadInfo(new LongIdKey(itemId), originFileName, content)
+            );
+
+            // 返回响应结果。
+            return FastJsonResponseData.of(ResponseDataUtil.good(null));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(Object.class, e, sem));
+        }
+    }
+
+    @PostMapping("/item-file/remove")
+    @BehaviorAnalyse
+    @BindingCheck
+    @LoginRequired
+    public FastJsonResponseData<Object> removeItemFile(
+            HttpServletRequest request, @RequestBody WebInputLongIdKey itemFileKey
+    ) {
+        try {
+            StringIdKey accountKey = tokenHandler.getAccountKey(request);
+            service.removeItemFile(accountKey, WebInputLongIdKey.toStackBean(itemFileKey));
+            return FastJsonResponseData.of(ResponseDataUtil.good(null));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(Object.class, e, sem));
+        }
+    }
+
+    private String adjustFileNameEncoding(String fileName) {
+        return new String(fileName.getBytes(), StandardCharsets.ISO_8859_1);
+    }
+}